Rename alerts saga
